+++ Source/EasyNetQ/DefaultTypeNameSerializer.cs
@@ -3,6 +3,7 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Reflection;
 using System.Text;
+using EasyNetQ.Internals;
 
 namespace EasyNetQ
 {
